# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kizuna Rail API",
    "version": "1.0.0",
    "description": "A brownfield API for CSE 341"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/api/schedules": {
      "get": {
        "summary": "Retrieve all schedules",
        "tags": [
          "Schedules"
        ],
        "responses": {
          "200": {
            "description": "A list of all schedules in the database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schedules": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Schedule"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch schedules"
          }
        }
      }
    },
    "/api/trips/{id}/schedules": {
      "get": {
        "summary": "Retrieve schedules for a specific trip (optionally filtered by month)",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The trip identifier (e.g., alpine-panorama)"
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional month value to filter schedule results"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of schedules matching the trip ID and optional month",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Schedules not found"
          },
          "500": {
            "description": "Failed to fetch schedules"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Schedule": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "6aad538fb32652d06b5deae0"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "tripId": {
            "type": "string",
            "example": "alpine-panorama"
          },
          "departureTime": {
            "type": "string",
            "example": "08:30"
          },
          "arrivalTime": {
            "type": "string",
            "example": "13:00"
          },
          "daysOfWeek": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday"
            ]
          },
          "status": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  },
  "tags": []
}